Sign conversion

+++ test/tests/libcxx/support/Emplaceable.h
@@ -44,7 +44,7 @@ template <>
 struct hash<Emplaceable>
     //: public std::unary_function<Emplaceable, std::size_t>
 {
-    std::size_t operator()(const Emplaceable& x) const {return x.get();}
+    std::size_t operator()(const Emplaceable& x) const {return static_cast<size_t>(x.get());}
 };
 
 }
